Windows Fix

+++ Registerer/src/Registerer/qrcode.py
@@ -5,6 +5,7 @@ erzeugt QR-Codes: muss dazu ausgedruckt werden
 liest QR-Codes
 '''
 import os, signal, subprocess
+import platform as p
 
 def erzeugenRawIn(strfile1="qrcode"):
     text=raw_input("Geben Sie einen Text für den QRCode an: ")
@@ -21,13 +22,17 @@ def erzeugen(text="", filename="qrcode"):
 
 
 def lesen():
-    zbarcam=subprocess.Popen("zbarcam --raw /dev/video0", stdout=subprocess.PIPE, shell=True, preexec_fn=os.setsid)
+    sys = p.system()
+    zbarcam=subprocess.Popen("zbarcam --raw /dev/video0", stdout=subprocess.PIPE, shell=True)
     print "zbarcam erfolgreich gestartet..."
     while True:
         qrcodetext=zbarcam.stdout.readline()
         if qrcodetext!="":
             print "success"
-            break
-    os.killpg(zbarcam.pid, signal.SIGTERM)  # Prozess stoppen
-    print "zbarcam erfolgreich gestoppt"
-    return qrcodetext
+            print "QR-Code: " + qrcodetext
+            if sys=="Windows":
+                subprocess.check_output("Taskkill /F /IM zbarcam.exe", shell=True).decode
+            else:
+                os.killpg(zbarcam.pid, signal.SIGTERM)  # Prozess stoppen
+            print "zbarcam erfolgreich gestoppt"
+            return qrcodetext
